+++ python/mqt/qmap/_compat/typing.py
@@ -11,11 +11,6 @@ from __future__ import annotations
 import sys
 from typing import TYPE_CHECKING
 
-if sys.version_info >= (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
-
 if sys.version_info >= (3, 11):
     from typing import Self
 elif TYPE_CHECKING:
@@ -24,7 +19,7 @@ else:
     Self = object
 
 
-__all__ = ["Self", "TypeAlias"]
+__all__ = ["Self"]
 
 
 def __dir__() -> list[str]: